[roll] Roll fuchsia [fxblob] Compress compression offsets in memory
Fxblob reads in multiple compressed chunks at once and ZSTD supports
decompressing multiple chunks at once. Only the compression offset of
of the first chunk of each read operation is needed, the other
compression offsets can be dropped. For 32KiB chunked blobs and a read
size of 128KiB, this will reduce the compression table size by 75%.

Compression offsets are stored as a u64 but a u32 would be sufficient
for all blobs that compress to less than or equal to 4GiB. This change
uses u32s to store all of the offsets less than or equal to u32::MAX
and u64s to store the rest. All of the current blobs are less than 4GiB
so this saves 50% of the offset table memory usage.

Original-Fixed: 398885530
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203846
Original-Revision: 9f8da6f4a646d8cdc4fc406bd2ae589e64d1e209
GitOrigin-RevId: 77f28aabf16b8e303ad42ea8ae5455c5a9a00c27
Change-Id: I170d814a4f355d9e8c4332720d8558b1481a937e