[roll] Roll fuchsia [zxdump] Handle ZX_INFO_PROCESS_MAPS with missing ZX_INFO_PROCESS_VMOS entry
This makes the zxdump memory dump logic more robust to getting a
ZX_INFO_PROCESS_VMOS sample that doesn't cover all the KOIDs that
appeared in the ZX_INFO_PROCESS_MAPS sample.  Heretofore this
would manifest as an assertion failure in a failed std::map::at
lookup.  Now it will just always elide the segment in question
from the dump as it would for a physical VMO or other "not safe
to attempt to dump" case.

It's suspected that this anomalous case is arising lately due to
a recent kernel regression.  But regardless, it's possible in two
ways for such an anomaly to arise without any bugs:
 1. Unsynchronized samples, either when dumping without
    suspending threads or when something outside the process is
    making mapping changes through a VMAR handle.
 2. Non-VMO mappings for iob and (soon) clock.

Original-Bug: 399510219
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216264
Original-Revision: 0efd2e3af1469aa0ac34e76dd252088b3011cd41
GitOrigin-RevId: 4e3ac6729013e90f80aeb773e0e3e2b0704fb503
Change-Id: Idd7aee7fa7928a7fbe69fb8c2c4212e285a4f862