[roll] Roll fuchsia [ffx][build] Removing unneeded ffx_plugin templates
This replaces ffx_plugin templates with rustc_libraries where the
ffx_plugin was not needed. The criteria is:
* no config data declared
* no plugin_deps
* compiled as a standalone subtool.

This is being done to remove complexity and confusion over the use
of the ffx_plugin and ffx_tool templates.

Original-Bug: 384505741
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216271
Original-Revision: a8c86bf639ddc99d0376da95141f6fab8e85b472
GitOrigin-RevId: 5bfd7633181b862c18ddb0185b200f64e25b3c43
Change-Id: I081c9eac78ecd92922c65034cbf260a36d573307