[roll] Roll fuchsia [gigaboot] shrink fastboot buffer to 1GiB
We previously allocated a 4GiB buffer for fastboot because we weren't
supporting sparse flashing, so we needed to be able to fit the multi-GiB
FVM image entirely in the buffer.

Now that we support sparse flashing we don't need such a large buffer,
and it was causing the emulator to fail to boot into fastboot since the
default emulator config only has 4GiB total.

Original-Bug: b/399567210
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216591
Original-Revision: 8cd2847072df6c5e8b216ca8eedd38d52b33f70a
GitOrigin-RevId: 802355039ef68f6f17bb792c4fbeeec467679258
Change-Id: I83f8e68802de64fde36798d542dedc6db1fbdab0