[roll] Roll fuchsia [fdomain] Wake up client task when there are handles to close
We try to batch the protocol messages that ask an FDomain to close a
handle to save us round trips, but we weren't flushing the batches soon
enough so things that depend on observing a handle closure might break.

A test for a subsequent patch exposed this error, which is why this
patch has no test itself.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1188196
Original-Revision: 65e4c64f3e5344d664c6882164bddc3e665d2e81
GitOrigin-RevId: 68d225a1b3153fae38aa16c46478136ef7b3b662
Change-Id: I1aadbad7de2ffc3ac9c1432e73d33599c0b88192