[roll] Roll fuchsia [starnix] Only return serialized bytes for SignalInfo to signals
For signals, one need to put the signalinfo struct associated with the
right ISA on the stack. This adds a method that always return an array
of the size of the 64 bit signal info, but filled with either the 32 or
64 bit version depending on the ISA.
This is used for putting the info on the stack, other callers should
always call write.

Original-Bug: b/395946855
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216707
Original-Revision: 9e4eccc6fa13d5cb76286a0f0a0c710dfbec34b6
GitOrigin-RevId: fe486649c23532296f7fecc833996b56f244f6c6
Change-Id: Ib2eb56b0f49e57bb2c68c90733b7a5976b1c5ec7