[roll] Roll fuchsia [ftdi-i2c] Don't forward metadata
Make the ftdi-i2c driver add its metadata when the `FtdiI2c` class is
constructed instead of when `DdkInit()` is called.

Originally, when the class was constructed it would create a device and
forward its metadata. Once `DdkInit()` was called it would replace the
forwarded metadata with new metadata. This made it pointless to forward
the metadata in the first place.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1215165
Original-Revision: 79d5d934c051b8a73ff4efab993485f077572c7a
GitOrigin-RevId: 1f2e070802c6b3060f665eeb26564e615ccccbba
Change-Id: I9b60ca749f5cdd9f0a40d5902fec23dab0664dc6