[roll] Roll fuchsia [component] Ensure services are opened as directories
Ensure that services are opened as directories instead of as protocol
connectors. For consistency with the GetServiceDirectory function, we
use the same set of permissions to ensure callers can enumerate protocol
instances within a service.

This is a precursor for migrating fdio_service_connect to use the new
Open method, as when we do, it would result in no permissions being
available on service instances.

Original-Bug: 324111518
Original-Bug: 376575307
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211864
Original-Revision: 23881ece53a615bdc611f4c22ed4a8cd9dae0419
GitOrigin-RevId: 7effa5ea78da677d80ab4b2692ab7b3b4d8112ff
Change-Id: Ic2ff6e917984c8b9446e029de9a7181d7c708fed