[roll] Roll fuchsia [fake-coordinator-connector] Rework package structure for tests.
This CL reduces the complexity of integration tests by switching to GN
templates that are or will be used by other parts of the display driver
stack.

This CL also clarifies the reasons for using the "vulkan" test type,
which is subtle and surprising to a reader who is not familiar with the
test types.

Original-Bug: 396160358
Test: fx test //src/graphics/display/testing/fake-coordinator-connector --count 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216331
Original-Revision: a10bd4504db70233d8b134006b90d991ae741ca1
GitOrigin-RevId: 6f028c2277fbf947787093d52e16dfc00578e4a7
Change-Id: I6c9295a7c648c278964e52d37cf84626694bf148