[roll] Roll fuchsia [starnix] Refactor empty children gating for unlink operations
This change helps mitigate the lock ordering bug described in
b/387576826. In short, the rename operation can potentially move
DirEntry nodes around the tree hierarchy. Since lock ordering assumes
a strict parent-to-child lock order, moving nodes around the tree can
result in false-positives for the mutex tracing graph.

The mitigation applied here is to offload the gating of non-empty
directories to the filesystem itself. This is arguably where the check
should belong in the first place, and has the added benefit of removing
the need for the DirEntry algorithm to obtain a lock on the children
of the node which we are unlinking.

This change adds a unittest to cover the above scenario. I've also
renamed "child" in the algorithm to instead be "child_to_unlink," which
I hope makes the code more readable.

Original-Bug: b/387576826
Test: Locally validated, added a unittest, and CQ.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206709
Original-Revision: bcfe56cb6cb16be4f1cb8f28fe0809aa8af80156
GitOrigin-RevId: 216e587760899e1e2a1e932810fc262112a2928c
Change-Id: Id95fb11ba4b6020bda51bd54877f41be574e2449