[roll] Roll fuchsia [display][fidl][banjo] Clarify engine / coordinator API docs.
This CL updates the comments on most calls in the
fuchsia.hardware.display.engine/Engine protocol. The main goal is to
clarify the API contracts, with an emphasis on preconditions and
postconditions.

This CL also removes duplicate comments from the Banjo version of the
protocol above. The comments are replaced with references to the
equivalent calls in the FIDL protocol.

Original-Bug: 316631158
Test: fx build (no code or API changes)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202144
Original-Revision: 20a2b7e2658fce7f4f2cb03c663ba109c632bdaa
GitOrigin-RevId: c081025cd0c0f7d6490cf2cc358bc45eae409add
Change-Id: I8d661e5a9b2d011540d1942f8aa44c730ab746a3