[roll] Roll fuchsia [idk][fidl] Unstable atoms only allowed in partner category
Require that FIDL libraries are stable unless they are in the "partner"
category by adding enforcement to `fidl()` similar to that already in
`sdk_source_set()` and `sdk_atom()`.

Also require `stable` to be specified when `sdk_category` is specified
and add it to the one library that didn't specify it.

In addition, simplify the logic enforcing stability in other categories
in `sdk_atom()`.

Original-Bug: 397525399
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210320
Original-Revision: 4b15d84daff61d68721a16bdded21b641c05fc74
GitOrigin-RevId: 810a16899cf7b13752b49d796acc146c8f47ee31
Change-Id: I7f8fd18334ae31b399d6b6153625eeadb0bd08f3