[roll] Roll fuchsia [sys] Ensure `OpenNamedServiceAt` uses Open
Ensure that services are opened as directories instead of as protocol
connectors. This was already done for `OpenNamedServiceIn` but was
missed for `OpenNamedServiceAt`.

This is a precursor for migrating fdio_service_connect to use the new
Open method, as when we do, it would result in no permissions being
available on service instances.

Original-Bug: 324111518
Original-Bug: 376575307
Test: fx test //sdk/lib/sys
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217584
Original-Revision: e5e8d031d55bfb678a07b4775f663aaff255b636
GitOrigin-RevId: c7dc60323849a6fc7385b1a6289248473f1eb11a
Change-Id: Id6922feaa4c632e94d3f85522c72b9611d3e90c7