[roll] Roll fuchsia [ffx][build] Remove sdk usage from mw-goldens
This reworks the mw-goldens tests to use ffx and the
subtool manifest instead of the SDK or the idk_atom_collection.

Also removes some unnecessary `async` function types.

And deletes a mis-located cli-golden file.

The change to the metrics code is to avoid loading the SDK to get the
SDK version when the analytics are disabled.

This improves the build performance since it no longer has to wait
for the IDK or SDK to be built.

Original-Bug: 392136182
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217963
Original-Revision: 00cf627277365b931d9d2e52af5ec7f01a5efb12
GitOrigin-RevId: 35bc1d2c0603ce100b25a8d7dfb7fb455d947cec
Change-Id: I957156d17ef0ffb2baf7b688e82c231780ed6fe9