[roll] Roll fuchsia [clidoc][build] Removing sdk-root option
This removes the use if the incorrectly named option "sdk-root"
this ends up being the build directory.

CLIDoc now uses a host tool atom collection to enumerate the tools
in included in the IDK, so there is no need for the IDK, or parts of
it, to be generated before calling clidoc.

Original-Bug: 392136182
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216998
Original-Revision: cd004eb500b1d8633e90bbdf0be5d02e95f0f812
GitOrigin-RevId: c30b9b44a9fc5cd02e7c71e6825a0819fcd2c2c7
Change-Id: I4d83360b501700a2eff2758d3c7cf26429574d10