[roll] Roll fuchsia [netstack3] Propagate more ICMP errors to TCP sockets
Currently there are a number of ICMP error codes that are not propagated
to TCP sockets, either as hard errors (which close the connection) or
soft errors (which don't close the connection but which can be retrieved
via SO_ERROR). Align the error propagation behavior with Linux, and add
integration test coverage for observing soft errors on established
connections, in addition to the existing coverage for hard errors on
attempted connections.

Original-Bug: 42052672
Test: netstack3-core-tcp-test
      netstack-socket-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216279
Original-Revision: cc1d67d2fa04157aa49bfd68077784d559f1a58e
GitOrigin-RevId: 3461e1a77ee70b2d15c8c909525c662add93dd74
Change-Id: If73980de3af61aec7885ba494457e2070d39ba0e