[roll] Roll fuchsia [kernel][arm64] Delay boot map init
This is a reland on http://fxrev.dev/1212556, which caused a crash in
start-up on sorrel. Previously the call was delayed until
arm64_mmu_early_init(), though it's unclear why that was an issue for
sorrel alone. In any case, it's clear that this call needs to be moved
up to before the periphmap is initialized (which is the last bit of code
directly dependent on the boot map)

We push back arm64_boot_map_init() to  platform_early_init(), just
before periphmap initialization. The boot map logic will go away
entirely soon, but until then there is no longer a need to call this in
assembly (as we're no longer setting up mappings in that context).

Original-Bug: 42164859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1218005
Original-Revision: 8dffc62871707ff0fae7d38a9e473f8fd7bd8f50
GitOrigin-RevId: 9b16688e5d940021faef906e4381b918135925b5
Change-Id: Icfcfe551efe4f3dc2d402f2a320cd61dad868cd2