[roll] Roll fuchsia [ffxutil] support ffx --strict emu start
This requires returning the "ffxInvoker", since the caller wants to
be able to both run the command, _and_ get the *exec.Cmd.

In strict, we require the emulator instance dir to be set, so we
now also add "emu.instance_dir" to the configs. And "ffx emu start"
specifically also requires "ssh.pub" to be set, so EmuStart() uses the
setConfigs() method to pass the additional config to the command.

Original-Bug: 391949505
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217630
Original-Revision: d352592176f9601b13c8034d3074b3c31fbc811a
GitOrigin-RevId: 0671fddb9178016a299b79345706ac33e861cadc
Change-Id: Ief0982d4289f0967aa5baba63b9863afb081835e