[roll] Roll fuchsia [idk][build] Build less for the in-tree IDK
Create separate IDK build graphs for the IDK used in-tree (by the Bazel
in-tree SDK and ffx tests) and the official IDK distributed to partners.
This decreases the build graph for most developers.

"//sdk:bazel_in_tree_idk" and "//sdk:sdk" now depend on the in-tree
graph. This avoids building unneeded targets when, for example,
assembling with the Bazel SDK.

"final_fuchsia_idk.exported" depends on the complete IDK graph, which is
now called "fuchsia".

The template named "idk_rules" implements the common logic for the two
graphs, which provide the specific targets to include when instantiating
the template.

Remove ":sdk_source_set_list" from "//sdk" since it is already a direct
dependency of "//:api" and would otherwise pull in the full IDK.

The same set of host tools are included in a default build, but tools
not needed as part of the SDK build are not included in the in-tree IDK.
This allows these tools to be built in parallel with targets that depend
on the IDK and SDK.

Also partially fix the dependency checking that triggers sub-build
rebuilds by adding a non-SDK deps in the main build on targets that are
produced for the IDK only in sub-builds. Packages are the remaining gap.

Original-Bug: 393154095
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1198649
Original-Revision: 00527f7a7c9582805fb5fa860266da56631e7060
GitOrigin-RevId: 4da86dffa563089b24633d32ce92e9a72abf8c19
Change-Id: I63f222ae9168e48de9186e4243a973d8429e846b