[roll] Roll fuchsia [tee] Hoist "ta" collection up to peer of ta-manager
This moves the "ta" collection to be a peer of the ta-manager component.
This matches the test setup. This makes it easier to set up test realms
as the capability for controlling the collection also controls
subpackage resolution. Now we can use the same ta-manager component in
production and tests and supply test TAs as subpackages in test realms.

This also moves the boilerplate capabilities and routing for hosting the
TA manager into a shard that we can use in tee.cml and (most) tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1220285
Original-Revision: 77ace85ef25e5bb5541cdfe2672e05990f9376bc
GitOrigin-RevId: cae9ab4c35631864208aad732e7aeb17cdb55288
Change-Id: Ic6b81123bad2a3258e9cb6389d14511a11694d9d