[roll] Roll fuchsia [kernel][iommu] New Map/Unmap IOMMU API
Changes the Map/Unmap design of the Iommu interface to use an opaque
mapping token, instead of operating directly on device address ranges.
This allows for separating the mapping of a range, with querying the
device contiguous portions of it. The existing IOMMU implementations
are able to support this change without actually needing to allocate or
store any additional metadata.

Original-Bug: 399697584
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216904
Original-Revision: 46d612f18b7931a3589010490d4f7e5a28fc7982
GitOrigin-RevId: 6efba8414814d7254f3e25b15bf1f68156a9104d
Change-Id: I32cd277748b0b1f9e5c4a5315d2dc6f40379c331