[roll] Roll fuchsia [dash] Update 'k' builtin to log an error if it can't send a debug command
Print an error message to stderr if this built-in is unable to send
a debug command to the kernel. When running this command from a non-serial
shell, such as a ssh shell, the command produces no direct output and so
failures are easy to miss.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221347
Original-Revision: 5e9046f96d132185caa8d6bd1ba54c5b1d10b0fb
GitOrigin-RevId: a6e703a40789f896fb911648ba8e2333d787101f
Change-Id: Ib67d43a78f32d54d1a9815f945b81f47d1f944fb