[roll] Roll fuchsia [fidl][rust] Relax static bound for transports
Driver framework uses an executor with a non-'static lifetime, and so
wants to be able to spawn futures with non-'static lifetimes as well.
This change pushes the requirement that transports be 'static down to
the site where futures are actually spawned. This can increase the
verbosity for more common cases, but it should be fine.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217464
Original-Revision: 5972bc9954da96a6e4100f0b1aae29b01aa056f8
GitOrigin-RevId: 60355743a6d42587aaabb215cb854f6477174fc1
Change-Id: I554aac6ebb484f1e5ce82c1a4d66fe633e9f65e5