[roll] Roll fuchsia [ui] Configure scene_manager with thread & memory profile.
On a heavily contended system it can a long time for scene_manager
to get scheduled and to resolve page faults when it needs to handle
input events. This will allow product integrators to tune its priority
relative to other programs contending for resources.

This should be a no-op on products which do not configure a role.

Original-Bug: 397790128
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221090
Original-Revision: 9652a18d0da0b5fca5492f4fee011fe8e4104dea
GitOrigin-RevId: f97a2d4696c0938f16b74e6d443416d2744ea68c
Change-Id: I318909e8d23892b1df766e75c89a8fb78a04df5b