[roll] Roll fuchsia [rbe][rust] Use innermost command for dep-scanning
In case the full rust command is wrapped with '--' command
prefixes, treat the last segment as the actual rustc command of
interest, used for deriving a dep-scanning a command.

This only fixes the first point of error in the referenced bug.

Original-Bug: 345596983
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222689
Original-Revision: 337ef107a880e9b538dc8ac0fa34458032d7ecf6
GitOrigin-RevId: 09082b9651004f2ed66eaf07195da1881c60aa19
Change-Id: I362122b6cf5657664dfe541cb969fbfb274fdd4b