[roll] Roll fuchsia [bazel_sdk] stop merging rules into fuchsia_sdk
This change removes the logic that merges the starlark rules into the
fuchsia_sdk repository. The users of the SDK have been migrated to the
new way of loading rules from rules_fuchsia.

There is still some cleanup work to be done with the build scripts to
remove the use_rules_fuchsia argument but that will be done in a later
CL since they are just setting that value to True everywhere and the
call sites for the generators have not been changed.

Original-Bug: 381104603
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222106
Original-Revision: 25e4527c09ab9a5253319d3fb5bfb467e22a796d
GitOrigin-RevId: ab0ce10a97dfff96d2a7cfaa312061b031caf972
Change-Id: I5ec4d1282936befa6e204bd69e5b8c0c0e3b4d74