[roll] Roll fuchsia [starnix][range_map] Implement RangeMap2
This CL contains a functioning copy-on-write implementation of RangeMap
that passes all the existing RangeMap tests. This code is an evolution
of the CowMap code, but specialized to implementing a range-based map.
Removing the abstraction between the RangeMap and the underlying Map
opens the door for future optimizations, such as removing the O(n^2)
gap search.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224745
Original-Revision: d01a04b0894ae24f4990c5ea2af10cbcbdcf3dc9
GitOrigin-RevId: 35f5317be589873663f90d6a98f0523fd7d9c8b0
Change-Id: Ie0c6c70f54632d45322b807ecf4fec081687c462