[roll] Roll fuchsia [netstack3] Simplify counter context traits
Remove the `with_` style methods from `CounterContext` and
`ResourceCounterContext`, replacing it with a flat method that returns a
reference to the underlying counters. This simplification makes it
possible for a follow up CL to access both "stack-wide" and
"per-resource" counters at once.

I suspect the original pattern was in place to facilitate putting
counters behind locks. However, that ability is not used by any of our
code today.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1225424
Original-Revision: 5784df777d31fc4af850298931a743afcff9eaeb
GitOrigin-RevId: 504102213913778593a80c0c3e3a0f33d6c02972
Change-Id: I8142efb0fef4854aaf6aade2077bd08cd78d9205