[roll] Roll fuchsia [owners] Remove jyotiraju@google.com from src/connectivity/telephony/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jyotiraju@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1160125
Original-Revision: eb8028b88695272c6469cc7ec18d5c2b0821394b
GitOrigin-RevId: 41e24a49f89465e469655b4653366222def1c71f
Change-Id: Id49b4b05282631dc61b9b73c7901764f9269989a