[roll] Roll fuchsia [netstack3] Introduce netstack3-trace
Replace TracingContext with a single crate that abstracts the fuchsia
tracing backend for netstack3 crates. This is simpler than the trait,
and allows us to use more of the facilities in tracing like appending
arguments without having to duplicate much of the API in the tracing
context.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1223247
Original-Revision: b88cb373f154fe9a259a4ff7efd08cb8ac2fd9d2
GitOrigin-RevId: deed09f9b97382a3df569ecd42d90a469f187a0f
Change-Id: I0dd9e10314a28e1e7b414ea4c78c18480f9f0b59