[roll] Roll fuchsia [build] Enable linker gc for Rust targets
This patch adds the necessary compiler and linker flags
to enable linker gc for Rust targets. There are some
symbols that are undefined in unused functions, and this
patch lets the compiler and linker to be able remove the
unused functions and avoid the undefined symbol error in
the linker.

Original-Bug: 402443020, 390250858
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1226604
Original-Revision: 27432643e8e971905cc992fe8dc8465b9e953127
GitOrigin-RevId: 0559dcb9f93c40c7e5599b690cf72ae49d3f6e4f
Change-Id: I82a59cb0a5fdb2ac1062425dcf2fa313b8825bc9