[roll] Roll fuchsia [kernel][vm] Free Loaned Pages Holder
Instead of moving loaned pages directly from a VMO->PMM they instead go
via a holding phase. While in this holding phase they are considered
immutable, and it is possible to wait for the page to cease being in
this temporary holding object.

An immutable holding object is necessary as the fine grained
VMO locking needs to be able to drop the lock to complete a range
change update, but a range change update must be completed before the
pages can be returned to the PMM.

Original-Bug: 338300943
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216964
Original-Revision: d39d22d531262813aaa4b87a3abb912d99484302
GitOrigin-RevId: b9c2b33e9b34de695907d052c5383b8e95286947
Change-Id: Icf60b64f9466353189c5e0c1280a98a70ea1a26e