[roll] Roll fuchsia [ffx][log] Add json schema support.
This adds json schema support for the logging command. It is done by
converting the main output from a `LogEntry` to a command-used
`CommandOutput` struct. The log entry is converted into this output for
simplicity's sake. The alternatives proved to be quite cumbersome (see
comments in the change).

Original-Bug: 392657374

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222014
Original-Revision: 3cf24f389d4acfc7aa03e876c7d75ee5451afd31
GitOrigin-RevId: e25fec81b4432fa46d7312768e5115d5b053a790
Change-Id: Ib8dcff3262f3ce469f6dedae445fa40496550d9c