[roll] Roll fuchsia [ffxutil] support strict config check-ssh-keys
Support ffx --strict config check-ssh-keys

This required a bit of a refactor, because the handling of ssh keys had
cyclical dependencies: some clients of FFXInstance wanted to provide
ssh key paths when constructing the instance, others wanted to query the
paths from the FFXInstance. There are also some clients which know both
the public and private key path, whereas others know only the public
one.

`ffx --strict config check-ssh-keys` requires both to be specified.
The current logic this requires the keys to be passed when in strict
mode (which is only used by botanist currently); when not in
strict mode, the keys are allowed to be determined from the environment.

Original-Bug: 400798715
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224748
Original-Revision: a8a05904208b906ea28eac3172159a31a94fb323
GitOrigin-RevId: 234148c2e5bcceb37590eedd5517488873536e07
Change-Id: Ieaf21fa6c4fc9b2ce020add79b53e812aaea7aa2