[roll] Roll fuchsia [bt][vendor] enable android vendor extensions on broadcom controllers
This change enables Sapphire to send the LE Get Vendor Capabilities HCI
command to the Controller on startup. Without this flag set, Sapphire
won't issue the command to the Controller: it’s unsafe to issue a
vendor-specific command and depend on the Controller to reject the
command if it doesn't understand it. The Controller could interpret the
command as a destructive operation, such as updating the Controller's
firmware. To safely implement vendor extensions and issue the related
HCI commands, we need to first publish support for a vendor-specific
command.

I've tested this command on smart displays, vim3, and various other
hardware that we support. The Controller responds properly and we
correctly handle the features that it supports.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217684
Original-Revision: 03f107692e9b2cb92cb33567a1ac00915f973b40
GitOrigin-RevId: ca7dcd848cd604c0820690cf095581e6532b0368
Change-Id: Ia2b3d42005f6f1e5ab830d1f11d69a816bda54ac