[roll] Roll fuchsia [libc] Move printf_core wrapper code out of the header
This starts a soft transition to enable llvm-libc to roll in with
changes to the printf_core internal API.  There's some temporary
conditionalization to handle either the old or new llvm-libc code
until the next roll lands.  This was facilitated by completely
separating the wrapper.h API from the actual llvm-libc code.  A
little more complete type erasure is done so that a separate file
provides the fully type-erased implementation.  Only that file,
and not everything using the header, needs to be exposed to the
llvm-libc internal headers directly.

Original-Bug: 402907255
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1227164
Original-Revision: 488bf6ec2e80148bb65a5efe8da202007523b573
GitOrigin-RevId: 19f22938af74baa83b8a27a174c9db24408e63e7
Change-Id: Ie11c21ec9a8579ee952fd41cff50c79863025451