[roll] Roll fuchsia [bazel_sdk] remove fuchsia_api_level alias
The fuchsia_api_level flags were moved to the fuchsia_sdk and thus we
needed to add an alias to enable a soft transition. All of the users of
the SDK have been updated so we this alias is no longer needed.

Original-Bug: 402169646
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228243
Original-Revision: e071696fcc01777011bdbae3b366b01d68cf1d99
GitOrigin-RevId: 0b863ffe04742e488d9dc55ca1e2747a972646b6
Change-Id: I722f0fd1514bba2430858e30e189cdd936bc1795