[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Fix new warning in DyadicFloat::as_mantissa_type_rounded (#131148)
The affected line of code converts a float's exponent from `int` to
`size_t`, negating it in the process. Following clang commit
773e88f9d61399c, this provokes a warning, presumably because the
conversion goes wrong if `size_t` is wider than `int` and the input
value is `INT_MIN`: negating it within the `int` type is undefined
behavior, with the likely (though not guaranteed) effect of leaving it
still at `INT_MIN` and then sign-extending that on promotion to
`size_t`.

This fix adds a cast so that the promotion to `size_t` happens _before_
the negation, so that the negative input value will _always_ be
sign-extended, and then the negation will make it positive again.

(I don't believe this case will actually come up. `DyadicFloat` is a
helper system used in processing ordinary float formats, so nothing is
expected to generate an exponent with even a 16-bit absolute value, let
alone 31. But it's as easy to fix it to be robust as it is to just
suppress the warning!)

GitOrigin-RevId: aa1cf9bb36ad95073ce9c52d7c8831fc4a5278d7
Original-Revision: 36e4f91294f0d061f4f33911378d919853aa626d
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228547
Original-Revision: bdbaa977ff69332dcd7be2b6e0a98490ef64c2df
Change-Id: I1e909b0a0741e19dddc9fdd285814e9e62016106