[roll] Roll fuchsia [vscode] Add trailing slash on remapPrefix
In some cases vscode seems to be inconsistent about whether or not the
${workspaceFolder} substitution has a trailing slash leading to things
only working intermittently or even not at all. This may have been made
worse some time recently as well.

This adds a trailing slash to the prefix so that it should always have
it, and seems to quell the errors in the log like:

```
File with cargo diagnostic not found in VFS: file not found: /usr/local/google/home/mgnb/fuchsiasrc/...
```

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1230464
Original-Revision: d7b1b14568780dd6c31addc05d171c4076dedd92
GitOrigin-RevId: 059e21c7540eda05e138c3e4ebac12fe8555ab4d
Change-Id: If377b1b5d83d02e4126689e49cb4ccaeacc9835d