[roll] Roll fuchsia [bazel_sdk] remove use_rules_fuchsia check in generator
As part of the migration to rules_fuchsia we added a way to soft transition the rules so not all SDKs needed to be updated. The soft transition is now complete so this CL removes the, no longer needed, soft transition logic.

Original-Bug: 381104603
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1230021
Original-Revision: b6ad0c550af320547105a3719d617d46cb165f56
GitOrigin-RevId: ff4b8dd9d176ff1da2c145d01483b7a6ea72ce06
Change-Id: I97032a305cf166069c59d78f486cb48c09eef7f3