[roll] Roll fuchsia [paver] Clone pager payload
Instead of attempting to assert that the payload VMO is not a root
pager VMO, this just directly creates a clone. Presently the check does
not work since the payload_vmo could be a reference clone, which is
both pager backed and has a parent koid, and so passes the check, but
since it still is the root VMO it could suffer from eviction.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228604
Original-Revision: 79aa84b35c592c2e2e0a71fa53d94777a13a7afd
GitOrigin-RevId: 407f7bc6c21f729d0984c701b32715284e8868db
Change-Id: I1c92826ce302e8f839f050a5e050e0dc66101e09