[roll] Roll fuchsia [brcmfmac] Adjust max TX glom frames
Recently the MTU in the driver was increased from 1500 to 1514, along
with that products are moving to ns3 which is better at utilizing frame
space. This resulted in the total size of a TX glom batch exceeding the
unsigned 16-bit limit in the glomming protocol header.

Adjust the maximum number of frames transmitted at a time to match the
maximum number of frames that can be TX glommed. Also add an error
message to indicate if the total size exceeds the maximum value. In the
future, this should ideally be improved to better recover if this
happens.

Original-Bug: 402604249
Test: Manual testing
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228148
Original-Revision: 32fc9774e3d8c27e665c91447225788ded74d298
GitOrigin-RevId: 01398125094993a4595d91bdab0e191e074cbcd3
Change-Id: I651eeaa208b0ae5c95b39e052674d270e3b69e2b