[roll] Roll fuchsia [zxio] Update zxio_create() on unknown type to successfully wrap
When zxio_create() is passed a valid handle to an object whose type it
does not understand it wraps the handle in a "holder" object that can
close the handle or provide it back to a caller. This allows zxio to
wrap any sort of handle in a zxio object. This change updates the return
value of this case to ZX_OK to make it easier for callers to understand
that the object is still valid. This handle holder behavior is useful
for transferring handles via zxio objects in some APIs - for example to
wrap an arbitrary handle in an fd to pass it through a binder protocol.

Previously this case returned ZX_ERR_NOT_SUPPORTED. This error case was
used during the fdio to zxio transition so that fdio could first ask
zxio to wrap a handle and if it was not something that zxio understood
fdio could then apply its own wrapping logic. That transition was
completed some time ago and so this confusing error handling is no
longer needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1231204
Original-Revision: c8bef371ec81e410a7c91d7e8bd337dd6615c636
GitOrigin-RevId: c4718ef10d395d0da920e5e14a45c6287b369bb0
Change-Id: I1a2bb2599c2c4d8b2c26aeef4881e39f8b32693c