[roll] Roll fuchsia [ffx][emu] Only generate a vsock CID if config value is 0
Right now we unconditionally overwrite whatever value is provided in the
config. While the user in most cases wouldn't care what value ends up
being used, it's strange to have this behaviour while exposing it as a
knob in the config. Accordingly, we update the default as 0 (an invalid
CID) and only overwrite it in that case. If any non-zero value is
provided, then that's the user intended and that's what the user will
get.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1227674
Original-Revision: 42bac5d1a3293446c474e321aa9ca870fb5d6eec
GitOrigin-RevId: 77e3488b9a5a6d88c2afdf43a8a41159b1c35860
Change-Id: I9a15c45a36336db78bc1ee44679a6c5866f259b7