[roll] Roll fuchsia [ffx] track remote overnet_id in the HostPipe
If we get a remote overnet id from the host-pipe, we'll need to
track it, and return it to callers. The specific plan is for
each TargetHandle object to request the overnet_id, and use that
to determine whether it should redirect its requests to a previously-
existing target.

Original-Bug: 395632395
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1229246
Original-Revision: 134b52125014279ce303c89a62a71166b1fd4982
GitOrigin-RevId: 227ec6908e24604cfe9c956197adc9b0d9f7728c
Change-Id: I7723eeb80e483859ecaf055284796065dbc35504