[roll] Roll fuchsia [ffx] reuse Targets when overnet node is shared
When a FIDL Target handler makes an RCS connection, check if it
is the same overnet id as one that already exists. If so, redirect
further Target requests to the previous target.

This required adding the ability to asynchronously wait for the
remote overnet id to arrive.

Original-Bug: 395632395
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1229247
Original-Revision: 44fd7192c792f62a13754d7d0d9769910b3d6ad2
GitOrigin-RevId: 338715fb7db1d916a230a17d763966bf667017cf
Change-Id: Ib040ad1e34f1d3eb2ca90db03018243325ef4a7e