[roll] Roll fuchsia [blobfs] Support for blob creator AllowExisting flag.
Supports AllowExisting flag so that a blob writer holds a reference to
a blob that it expects to overwrite. During commit it replaces the
loader info for the old blob with the new one and commits the old blob's
removal to the journal in the same transaction.

If the disk writes fail such that the transaction cannot be completed,
then the blob is marked as corrupt to prevent attempts to load it
afterward as the metadata and/or block data for the blob is in an
unknown state.

Original-Bug: 400777068

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1229252
Original-Revision: f014f5ed101dbfd70850714bd92ed2f5480cb50a
GitOrigin-RevId: 07bad058e0a26a7932e2e17e57513c3b2b7b4eff
Change-Id: I78680993c4f17b0164b3d7722e2f705a9afa38ae