[roll] Roll fuchsia [ffx][trace] Replace a mutable out-arg
As pointed out previously, it's kind of gross. I think I've been
spending too much time writing C++ lately.

Instead of having the symbolize_event_record take a Vec<u8> and &mut
bool and returning a Vec<u8> which is either the same as what was passed
in or a modified one (in which case the out-arg would be set to true, it
takes a &[u8] and returns an Option<Vec<u8>> so that the caller can
determine if there's been any modification.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233105
Original-Revision: b43545c247b8bf05f3fb4c07bc6ec6c31f3b4ab0
GitOrigin-RevId: ae2640f9390399b9e7cb00bf652c87544941c306
Change-Id: I6839065e3120d43f94800ebdecda3f91174007fa