[roll] Roll fuchsia [starnix][vfs] Create FsNodeRareData struct
The FsNode structure contains many fields that are very rarely used.
This CL creates an FsNodeRareData struct to hold those fields.

In a typical usage sceniario we have about 7 MiB of FsNodes, of which at
least 2.5 MiB are the WaitQueues inside if `flock_info` and
`record_locks`. This CL probably saves more memory than that, but it's
just easy to see those fields in the heap profile because they are
separate allocations.

This CL removes 85% of the WaitQueues in the entire Starnix kernel.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233008
Original-Revision: 24f55ae701df6e4231a2362328482f518c079fcb
GitOrigin-RevId: 5c64e47200dd100c3c8fb6184f70592aa8fd71de
Change-Id: Ia412fc6c18180dd8c9a90a247dff205ee087b531