[roll] Roll fuchsia [netstack3] Extract SeqRanges data struct
Extract the core of the Assembler struct into SeqRanges so we can reuse
it later for the SACK scoreboard. The core of this is to keep track of
disjoint ranges in an ordered fashion and this is symetrical for the
send side scoreboard.

Original-Bug: 42078221
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233565
Original-Revision: 1f3cff48a40a3adda07f97b88e4b2251739b9802
GitOrigin-RevId: 4b7febcc6fba3eee9793d13385cd26c673879c7d
Change-Id: I512503e0c60d2af965291ceb5dd8dddbcba777c5