[roll] Roll fuchsia [assembly][bt] Refactor the Audio Gateway config
Define a new schema for HFP AG that uses Rust enums to apply additional
features only when the profile is enabled. This matches the schema for
HFP Hands Free.

The new schema disambiguates between enabling Hands Free and Audio
Gateway and allows for the enabling of both.

```
"profiles": {
  "hfp": {
    "audio_gateway": {
      "feature1": true,
      "feature2": true,
      ...
    },
    "hands_free": {
      "feature1": true,
      ...
    },
  }
}

The supported codecs and controller encoding schemes are common to both
profiles and therefore kept in the toplevel HfpConfig.

Define a new AudioGatewayConfig which is used to enable the profile with
any features. Since some products do enable this via the `enabled` field,
keep it in the HfpConfig until all products are migrated to the new
schema. All other existing fields are not specified in any products and
are directly  moved to the AudioGatewayEnabled config.

Original-Bug: 401064356
Test: fx build
Test: fx test src/lib/assembly
Test: Manual, fx set smart*, fx build to verify this is backwards
compatible

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222204
Original-Revision: 7a417d6b8a15a76db11547a361b12f0e22767e77
GitOrigin-RevId: 7710acf591ac2178eba59d9d6307725772c09854
Change-Id: Ia6312a4f5d40d20188ae7fca073ffc6d7cfceac7