[roll] Roll fuchsia [dl] RuntimeModule has name ownership
This CL changes the `name_` field for RuntimeModule into an owned char
array that will convert to an Soname object through the RuntimeModule
API. This char array is guaranteed to contain a nul byte.

A test is added to test that the caller of dlopen can garbage collect
the filename pointer after a RuntimeModule has been created for that
file.

Original-Fixed: 335921712
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233106
Original-Revision: e576f4bce4ad48e52dcd472d08a9faefd630ec86
GitOrigin-RevId: 632dfd0addcafba9aa1da521c8f6c737e479543e
Change-Id: Icd0a2c6a5d7293f30e30760798c3b7d1c43713db