[roll] Roll fuchsia [ffx][sdk] Adding SdkRoot::HostTools
This sdk root is the "root of last resort", when no SDK manifest can
be found, and the sdk.root is not explicitly set.

This commonly happens in integration tests in-tree that are using `ffx`
and an isolate dir.

The behavior of the SDK when using and HostTools SDKRoot is that
get_host_tool() will look in the same directory as the current process,
which is most likely ffx).

Adding this SdkRoot type enables running integration tests without
having to build the entire IDK, which will make iterating on these
tests faster.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232964
Original-Revision: 7fba14a7ba0bbd8439ca269cb0cd4d84b0adc5c0
GitOrigin-RevId: cb6c2ad8e8de8937f6b18dfdfb52e4e8785d92bc
Change-Id: I657d6972a2ab1f647795a7869206f73d11ec95aa