[roll] Roll fuchsia [storage] Use fuchsia_sync everywhere
Replace std::sync types with the equivalent fuchsia_sync types. Fuchsia
builds with panic=abort so we don't need to handle Mutex poisoning so we
always unwrap the result of locking a Mutex. fuchsia_sync doesn't do
Mutex poisoning and doesn't return a Result when locking a Mutex.
fuchsia_sync also provides MappedMutexGuards that are still experimental
in std::sync.

This is not expected to have an impact on performance or memory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1231904
Original-Revision: 9a2b17841b5a937496fbf0dbd751a04641b66ca7
GitOrigin-RevId: 2160a860f4c60962eab18123e30787369928dfb9
Change-Id: I638c98d0499b8b73b0a5e67eca263cb276da35d3