[roll] Roll fuchsia [wlan][policy] Match roam request to pending roam
This change adds a way to track pending roams initiated by
the policy layer, and verify that we receive a matching roam result. If
a different roam result is received, or if we hit a timeout without
receiving a matching result, something has gone wrong and the connection
must be terminated. Temporary solution for b/399649753.

Original-Bug: 399649753
Test: fx test wlancfg-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211948
Original-Revision: c53b0a219eccc444727e83faa56006d7ea27c172
GitOrigin-RevId: 239f9708c52dad16c36961834bd3c63cb9f776d9
Change-Id: I3516e9ab4f0ac9f7e0fb8ec773a798d69777abba