[roll] Roll fuchsia [kernel] Rename KTrace::GetInstance -> GetInternalState
The GetInstance function was actually returning a reference to the
internal::KTraceState, so rename the function as such. This will
reduce confusion when we create a static KTrace instance, which we
will need to support the new per-CPU buffer architecture.

Original-Bug: 404539312
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235926
Original-Revision: edf3da3040bf9328c938f675c3c4646e7ff0184f
GitOrigin-RevId: 880181edd7c28ff523f5a63c7d7c1094b42cfeab
Change-Id: I73eba0d01c32ec3e5f5bf2777b18984a2b16388a