[roll] Roll fuchsia [fshost] Verify open success during migration recursive copy.
Without verifying this we can race in the tests because the file or
directory may not actually be created by the time we verify it. More
than that, src may be empty which means that we never interact with
the dst copy again and never verify that it was successfully created
before destroying the old filesystem.

Original-Bug: 405230091
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1234224
Original-Revision: 35d6b311ecd1683a9d2210639f3a65acc738c93c
GitOrigin-RevId: 6e48fb9b17c349aaf7d8088563002050d6313168
Change-Id: I0ca5199f6219a26ea97d34e50edb48422f5ed386