[roll] Roll fuchsia [starnix] Containers have a structured config schema.
This change defines a set of configuration capabilities in
//src/starnix/containers/container_config.shard.cml which can be
offered to any Starnix container to configure its behavior. These
configuration capabilities can extend or override the container
definition in a container's package.

For an initial example, makes it possible to add extra feature strings
for a particular instantiation of a container. This will allow tests to
keep a shared container definition but to turn on particular features
of that container which are not needed by all tests that use the
container.

The syscall tests generate an additional component for each suite that
exposes the configuration values. In many cases it is possible to
define config values in the same component that offers the capabilities,
but here that would create a dependency cycle since the test component
also `use`s capabilities from the container.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235174
Original-Revision: 6e9e865eed003239b28409d0a819e542a8d5a1d3
GitOrigin-RevId: 2683ff9c2d603616bda4dabb7b95d590cbfc9690
Change-Id: I55b2fe851fd0356e85aff1b9762536eb4f6bc1d4