[roll] Roll fuchsia [assembly] Support AB slotted recoveries
If a product has A and B slots for recovery, the partitions config can
be updated to indicate this. When flashing the images, the recovery
zbi/vbmeta will be flashed to the A and B slots rather than the R slots.

To indicate this to assembly use:

  fuchsia_partition(
      name = "recovery_a",
      partition_name = "recovery_a",
      slot = SLOT.A,
      type = PARTITION_TYPE.RECOVERY_ZBI,
  )

Instead of:

  fuchsia_partition(
      name = "recovery_r",
      partition_name = "recovery_r",
      slot = SLOT.R,
      type = PARTITION_TYPE.ZBI,
  )

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233751
Original-Revision: da251c051ca3b6c45801889ae86be912b0ddbf72
GitOrigin-RevId: 292774a105aa9f49e448de9f1dd5c96323888dbf
Change-Id: Idfa433a96bb752f76debe522d5132e7f2d7a881b