[roll] Roll fuchsia [flatland] Fix buffer collection image import
Upon "reenabling" direct-to-display, I discovered that it wasn't
actually working, because the current code required the Flatland client
to maintain a reference to the imported buffer collection for as long as
the the image was being used to render. At a glance, this behavior seems
to date back to 2022, but this conflicts with more recent observations.

The fix is to validate the image against `display_imported_images_`.

Also, a couple cleanups to more tightly scope local variables.

Original-Fixed: 377979329
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232464
Original-Revision: 0822c3663e37a44f71f27a31fc19f9841389f351
GitOrigin-RevId: 6798bb7ca62bbd5e20868278981363c8873b1275
Change-Id: I2c9c5539a060da15142afee670cf9f850c846c41