[roll] Roll fuchsia [third-party][rust] Generate BUILD.bazel
Initial generation of BUILD.bazel files for third-party Rust crates.
This CL is generated with:

> fx update-rustc-third-party

NOTE: It is not guaranteed that __all__ of the generated targets are
correct. I only validated a few of them by building
`//build/tools/formatjson5` in Bazel. In fact, all the targets that
have extra cargo-gnaw annotations in Cargo.toml won't build in Bazel.
They will be fixed with matching crate universe annotations in followup
changes on a case-by-case basis.

Original-Bug: 358308332

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232644
Original-Revision: 82f08267ef0a674341ebaac0469cbe2134064426
GitOrigin-RevId: 2b3c5a3943dcbc5acae5d17a038be9822c34a3b7
Change-Id: I953d37fab56d26091ad19898f909d89ea5fa7203