[roll] Roll infra/recipes [sdk] Skip nonexistent branches when resolving CIPD refs
We intentionally include a nonexistent branch corresopnding to the next
milestone in the `ref_settings` for canary SDK builds so that the
settings will be correct once the branch is created with no need for
cherry-picks. This was fine up until now because the only builds that
triggered with nonexistent branches in `ref_settings` were on the canary
branch, and the canary branch just happened to be listed first in
`ref_settings`.

However, now we've added another branch at the end of `ref_settings`
that triggers builds, so when passing over the next milestone branch we
would fail to process it, for example:
https://ci.chromium.org/b/8719311835597979057

So make sure not to fail if the branch doesn't exist, and instead
proceed to check other branches.

Original-Bug: 406038575
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1236209
Original-Revision: 8dd189ddad76e150d25e228adb4b96983aaeba02
GitOrigin-RevId: 1ca63f043cdb17753b68fe294f8803a667700211
Change-Id: I435c8fc2fcde87e4e086aa94d1336e1b38c1e474